Provide events with XML chat data
