+++ src/Modules/WEBSERVER_MODULE/CommandReplyEvent.php
@@ -7,7 +7,7 @@ use Nadybot\Core\Event;
 class CommandReplyEvent extends Event {
 	/**
 	 * An array with reply messages
-	 * @var string[]
+	 * @var AOMsg[]
 	 */
 	public array $msgs = [];
